refactor(core): split_args を snotra-core へ移設しオープナーと共有
