+++ snotra-core/src/instant.rs
@@ -2,6 +2,31 @@ use percent_encoding::{NON_ALPHANUMERIC, utf8_percent_encode};
 
 use crate::config::InstantCommand;
 
+/// シェル風クォート対応の引数分割。
+/// `"..."` で囲まれた部分はスペースを含んでも1トークンとして扱う。
+/// 閉じクォートがない場合は行末まで1トークン。
+/// 空クォート `""` はトークンを生成しない。
+pub fn split_args(args: &str) -> Vec<String> {
+    let mut tokens = Vec::new();
+    let mut current = String::new();
+    let mut in_quotes = false;
+    for ch in args.chars() {
+        match ch {
+            '"' => in_quotes = !in_quotes,
+            c if c.is_whitespace() && !in_quotes => {
+                if !current.is_empty() {
+                    tokens.push(std::mem::take(&mut current));
+                }
+            }
+            c => current.push(c),
+        }
+    }
+    if !current.is_empty() {
+        tokens.push(current);
+    }
+    tokens
+}
+
 /// Expand `{query}` and `{clip}` placeholders in an instant command template.
 ///
 /// If the command starts with `http://` or `https://`, variable values are
@@ -166,5 +191,27 @@ mod tests {
         ];
         let result = filter_instant_commands(&cmds, "google");
         assert_eq!(result.len(), 1);
+    }
+
+    // ---- split_args (quote-aware splitting) tests ----
+    #[test]
+    fn split_args_quoted_token_preserves_spaces() {
+        assert_eq!(split_args(r#"--dir "My Documents""#), vec!["--dir", "My Documents"]);
+    }
+    #[test]
+    fn split_args_unclosed_quote_consumes_to_end() {
+        assert_eq!(split_args(r#"--dir "My Documents"#), vec!["--dir", "My Documents"]);
+    }
+    #[test]
+    fn split_args_adjacent_quotes_join() {
+        assert_eq!(split_args(r#"--open="My File""#), vec!["--open=My File"]);
+    }
+    #[test]
+    fn split_args_empty_quotes_produce_no_token() {
+        assert_eq!(split_args(r#"a "" b"#), vec!["a", "b"]);
+    }
+    #[test]
+    fn split_args_plain_whitespace_only() {
+        assert_eq!(split_args("  -a   -b  "), vec!["-a", "-b"]);
     }
 }
