+++ src-tauri/src/commands/launch.rs
@@ -2,6 +2,7 @@ use std::time::Duration;
 
 use serde_json::json;
 use snotra_core::config::{find_matching_tools, OpenerTool};
+use snotra_core::instant::split_args;
 use tauri::{AppHandle, Manager};
 use tokio::time::timeout;
 
@@ -171,34 +172,6 @@ fn build_launch_args(args: &str, path: &str) -> Vec<String> {
     expanded
 }
 
-/// シェル風クォート対応の引数分割。
-/// `"..."` で囲まれた部分はスペースを含んでも1トークンとして扱う。
-/// 閉じクォートがない場合は行末まで1トークン。
-/// 空クォート `""` はトークンを生成しない（空の引数を明示渡しする手段は提供しない）。
-fn split_args(args: &str) -> Vec<String> {
-    let mut tokens = Vec::new();
-    let mut current = String::new();
-    let mut in_quotes = false;
-    for ch in args.chars() {
-        match ch {
-            '"' => {
-                in_quotes = !in_quotes;
-            }
-            c if c.is_whitespace() && !in_quotes => {
-                if !current.is_empty() {
-                    tokens.push(std::mem::take(&mut current));
-                }
-            }
-            c => {
-                current.push(c);
-            }
-        }
-    }
-    if !current.is_empty() {
-        tokens.push(current);
-    }
-    tokens
-}
 
 #[tauri::command]
 pub async fn launch_item(
@@ -415,44 +388,6 @@ mod tests {
             build_launch_args("{path} --compare {path}", "C:\\file.txt"),
             vec!["C:\\file.txt", "--compare", "C:\\file.txt"]
         );
-    }
-
-    // ---- split_args (quote-aware splitting) tests ----
-
-    use super::split_args;
-
-    #[test]
-    fn split_args_quoted_token_preserves_spaces() {
-        assert_eq!(
-            split_args(r#"--dir "My Documents""#),
-            vec!["--dir", "My Documents"]
-        );
-    }
-
-    #[test]
-    fn split_args_unclosed_quote_consumes_to_end() {
-        assert_eq!(
-            split_args(r#"--dir "My Documents"#),
-            vec!["--dir", "My Documents"]
-        );
-    }
-
-    #[test]
-    fn split_args_adjacent_quotes_join() {
-        assert_eq!(
-            split_args(r#"--open="My File""#),
-            vec!["--open=My File"]
-        );
-    }
-
-    #[test]
-    fn split_args_empty_quotes_produce_no_token() {
-        assert_eq!(split_args(r#"a "" b"#), vec!["a", "b"]);
-    }
-
-    #[test]
-    fn split_args_plain_whitespace_only() {
-        assert_eq!(split_args("  -a   -b  "), vec!["-a", "-b"]);
     }
 
     #[test]
